requirements.txt
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
databento>=0.36.0
requests>=2.31.0
